Rename create_doc_checklist in metadata_checklist
