+++ napari_hub_cli/documentation_checklist/metadata_checklist.py
