+++ tests/test_doc_checklist_repo1.py
@@ -1,13 +1,13 @@
 from pathlib import Path
 
 import pytest
-from napari_hub_cli.documentation_checklist.create_doc_checklist import (
+
+from napari_hub_cli.documentation_checklist.filesaccess import NapariPlugin
+from napari_hub_cli.documentation_checklist.metadata_checklist import (
     DISPLAY_NAME,
     Feature,
     check_feature,
 )
-
-from napari_hub_cli.documentation_checklist.filesaccess import NapariPlugin
 
 
 @pytest.fixture(scope="module")
